### Motivation
- Fix a Dart analyzer warning about an unawaited Future started by the
lyrics line transition and make the animation intent explicit to avoid
dropped futures and spurious warnings.
- Prevent unnecessary Media3 media-item replacements and potential
subtle bugs by comparing artwork byte arrays by content instead of by
reference.
- Eliminate unsafe nullable assumptions and minor Kotlin warnings in
palette/metadata/notification code to improve robustness and avoid
runtime issues on some devices.

### Description
- Marked the synced-lyrics animation call as intentionally unawaited:
`unawaited(_lineTransitionController.forward(from: 0.0));` in
`lib/widgets/player/synced_lyrics_view/state_indexing.dart`.
- Compare artwork payloads by content with `contentEquals` to avoid
redundant `replaceMediaItems` in
`android/app/src/main/kotlin/dev/wndavenz/music/Media3PlaybackService.kt`.
- Removed unsafe non-null assertions and added fallbacks in palette
selection in `NativePaletteBridge.kt` so secondary/accent clusters are
handled safely and fallback derivation is preserved.
- Cleaned a few Kotlin expressions that could cause warnings or
misbehavior: simplified frequency conversion in
`AudioEffectsManager.kt`, trimmed/comment parsing in `TagBuilder.kt`,
and tightened notification cache checks in
`PlaybackNotificationManager.kt`.
- Applied Dart formatting fixes and small white-space/line-break
normalizations in several Dart files including `state_build.dart` and
settings widget parts.

### Testing
- Ran `dart format lib` which completed and formatted files (formatted
278 files, several small changes applied). — succeeded.
- Ran `flutter analyze` on the project and it reported `No issues
found`. — succeeded.
- Executed `flutter test` (Dart unit tests) and all Dart tests passed. —
succeeded.
- Built a release APK with `flutter build apk --release` which produced
`build/app/outputs/flutter-apk/app-release.apk` (28.3MB). — succeeded.
- Ran Android unit tests `./gradlew :app:testDebugUnitTest` and
`./gradlew :app:testDebugUnitTest` succeeded for the app module unit
tests. — succeeded.
- Ran native package tests `cd native_audio_runtime && dart test` and
they all passed. — succeeded.
- Executed the broader Gradle test task `./gradlew test` which runs
transitive plugin Robolectric tests; some plugin tests
(`url_launcher_android`, `shared_preferences_android`) failed due to
external Maven/Robolectric artifact fetch errors caused by
`java.net.SocketException: Network is unreachable`, not by the code
changes. — partially failed (network-related, not code regressions).

------
[Codex
Task](https://chatgpt.com/codex/cloud/tasks/task_e_6a7a120ccaf883328b3504064de406c8)

+++ android/app/src/main/kotlin/dev/wndavenz/music/effects/AudioEffectsManager.kt
@@ -293,7 +293,7 @@ class AudioEffectsManager(private val effectHandler: Handler) {
                 // This is the actual hardware-reported center frequency for each band,
                 // which varies by device and Android version.
                 "frequencies" to List(numBands) { b ->
-                    (eq.getCenterFreq(b.toShort()) / 1000).toInt()
+                    eq.getCenterFreq(b.toShort()) / 1000
                 }
             )
         } catch (_: Exception) { defaultEqualizerParameters() }
